feat: refactoring create_ingress to kubernetes_manager module
