+++ app/v1/modules/kubernetes_manager/__init__.py
@@ -3,6 +3,7 @@
 
 from .script_create_configmap_or_secret import ScriptConfigMapSecretCreator
 from .script_create_environment_values_file import ScriptEnvironValuesFileCreator
+from .script_create_ingress import ScriptIngressCreator
 from .script_create_vars_file import ScriptVarsFileCreator
 from .script_create_workload_template import ScriptWorkloadTemplateCreator
 from .script_create_workload import ScriptWorkloadCreator
@@ -11,6 +12,7 @@ from .script_create_workload import ScriptWorkloadCreator
 __all__ = [
     "ScriptConfigMapSecretCreator",
     "ScriptEnvironValuesFileCreator",
+    "ScriptIngressCreator",
     "ScriptVarsFileCreator",
     "ScriptWorkloadTemplateCreator",
     "ScriptWorkloadCreator",
